feat(discord/create-new-scheduled-event): rework fields
