+++ packages/backend/src/apps/discord/dynamic-fields/index.ts
@@ -1,0 +1,3 @@
+import listExternalScheduledEventFields from './list-external-scheduled-event-fields';
+
+export default [listExternalScheduledEventFields];
