+++ packages/docs/pages/apps/discord/actions.md
@@ -3,8 +3,8 @@ favicon: /favicons/discord.svg
 items:
   - name: Send a message to channel
     desc: Sends a message to a specific channel you specify.
-  - name: Create a new event
-    desc: Creates a new server event.
+  - name: Create a new scheduled event
+    desc: Creates a new scheduled event.
 ---
 
 <script setup>
